requirements.txt
llama-cpp-python>=0.2.11
pyyaml>=6.0
requests>=2.28.0
pytest>=7.0.0